+++ src/models/ListPracticesResponseDataInner.ts
@@ -126,7 +126,7 @@ export interface ListPracticesResponseDataInner {
    */
   supportPhone: string | null;
   /**
-   * Match this integration's external identity in the API key's mode.
+   *
    * @type {string}
    * @memberof ListPracticesResponseDataInner
    */